% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_physchem.R
\name{get_physchem}
\alias{get_physchem}
\title{Retrieve Physio-Chemical Properties}
\usage{
get_physchem(type, value)
}
\arguments{
\item{type}{a character string indicating}

\item{value}{the value for the matching \code{type}}
}
\value{
a \code{tibble} containing the following fields for each returned entry
\itemize{
\item \strong{lm_id}
\item \strong{molecular_weight}
\item \strong{exact_mass}
\item \strong{heavy_atoms}
\item \strong{rings}
\item \strong{molecular_volume}
\item \strong{rotatable_bonds}
\item \strong{hbond_donors}
\item \strong{hbond_acceptors}
\item \strong{slogp}
\item \strong{smr}
\item \strong{tpsa}
\item \strong{aromatic_rings}
}
}
\description{
Retrieve Physio-Chemical Properties
}
