% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMsearch.R
\name{LMsearch}
\alias{LMsearch}
\title{Seach LIPID MAPS}
\usage{
LMsearch(x)
}
\arguments{
\item{x}{a character of a valid \code{InChiKey} or \code{Molecular Formula}}
}
\value{
a \code{data.frame} containing the following fields for each returned entry
    \describe{
        \item{name}{Systematic name}
        \item{id}{LIPID MAPS database id}
        \item{mz}{Monoisotopic accurate mass}
        \item{inchi}{InChi Code}
        \item{inchi_key}{InChiKey}
        \item{smiles}{Smiles}
        \item{formula}{Molecular formula}
        \item{core_class}{core}
        \item{main_class}{main}
        \item{sub_clss}{sub}
    }
}
\description{
Retrieve data from the LIPID MAPS database using either an \code{InChiKey} or \code{Molecular Formula} as an
input
}
\examples{
\dontrun{
    LMsearch("C18H32O2")
    LMsearch("ZKRPGPZHULJLKJ-JHRQRACZSA-N")
}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

