% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_by_mass.R
\name{search_by_mass}
\alias{search_by_mass}
\title{Search the LIPID MASS Virtual Database}
\usage{
search_by_mass(mz, adduct, ppm = 1)
}
\description{
Perform a precursor (MS1) ion search using the LIPID MAPS virtual database
}
