% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_lipidmaps.R
\name{search_lipidmaps}
\alias{search_lipidmaps}
\title{Search LIPID MAPS}
\usage{
search_lipidmaps(type, value)
}
\arguments{
\item{type}{a character string indicating}

\item{value}{the value for the matching \code{type}}
}
\value{
a \code{tibble} containing the following fields for each returned entry
- \strong{name} Systematic name
- \strong{id} LIPID MAPS database id
- \strong{mz} Monoisotopic accurate mass
- \strong{inchi} InChi Code
- \strong{inchi_key} InChiKey
- \strong{smiles} Smiles
- \strong{formula} Molecular formula
- \strong{core_class}
- \strong{main_class}
- \strong{sub_class}
}
\description{
Search the LIPID MAPS database
}
\examples{

search_lipidmaps('inchi_key', 'ZQPPMHVWECSIRJ-KTKRTIGZSA-N')

search_lipidmaps('formula', 'C20H34O')

search_lipidmaps('pubchem_cid', 445639)
}
